+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Yii Error Handler Change Log
 
-## 3.0.1 under development
+## 3.1.0 under development
 
-- no changes in this release.
+- Chg #75: Dispatch `ApplicationError` in `ErrorCatcher` (@xepozz)
+- Enh #82: Add `HeadersProvider` (@xepozz)
 
 ## 3.0.0 February 14, 2023